android_test_automation/requirements.txt
# Android Test Automation Requirements
# AI Insect Bug Identifier App Testing

# Appium for Android automation
Appium-Python-Client>=3.1.0
selenium>=4.15.0

# Image processing (for augmentation integration)
opencv-python-headless>=4.7.0
pillow>=9.4.0
numpy>=1.23.0
albumentations>=1.3.0

# Utilities
python-dateutil>=2.8.2

# Optional: For Excel support (if needed)
# openpyxl>=3.1.0
# pandas>=2.0.0




